requirements.txt
mss
Pillow
pynput
pystray
pywin32; sys_platform == "win32"
